Add attributes

+++ database/migrations/2017_09_10_114031_category.php
@@ -13,22 +13,26 @@ class Category extends Migration
      */
     public function up()
     {
-        Schema::create('categories', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('parent_id')->nullable()->default(NULL);
-            $table->string('name');
-            $table->string('slug');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('categories')) {
+            Schema::create('categories', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('parent_id')->nullable()->default(NULL);
+                $table->string('name');
+                $table->string('slug');
+                $table->timestamps();
+            });
+        }
 
-        Schema::create('category_product', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('category_id')->unsigned();
-            $table->integer('product_id')->unsigned();
-            $table->timestamps();
+        if (!Schema::hasTable('category_product')) {
+            Schema::create('category_product', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('category_id')->unsigned();
+                $table->integer('product_id')->unsigned();
+                $table->timestamps();
 
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
-        });
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
+            });
+        }
     }
 
     /**
